Continue to abstract out commonalities and unnecessaries
